refactored the exercise

+++ src/shapes/Rectangle.java
@@ -1,21 +1,21 @@
 package shapes;
 
 public class Rectangle {
-    protected int length;
-    protected int width;
+    protected double length;
+    protected double width;
 
     // constructor
-    public Rectangle(int length, int width) {
+    public Rectangle(double length, double width) {
         this.length = length;
         this.width = width;
     }
 
     // methods
-    public int getArea() {
+    public double getArea() {
         return this.length * this.width;
     }
 
-    public int getPerimeter() {
-        return 2 * this.length + 2 * this.width;
+    public double getPerimeter() {
+        return (2 * this.length) + (2 * this.width);
     }
 }
